Making oscar contrib models abstract
Splitting the oscar model content block into concrete and abstract
varieties to allow for easier extension.
